daemon: move purchase `jsonrpc_` methods to an individual module
